Move `TestRunnerFormula` class into a separate file
